* arts-1-11-27
* Lots of casting fixes to make account for the stricter warning
policy in gcc 4.3. Warnings are now issued for all implicit conversions
that might lead to alteration of values. Like casting from Index to
Numeric. The fix is to insert explicit castings in the corresponding
places to show the compiler that we know what we're doing and are
aware of the problem.

* configure.in: Reenable -Werror for gcc 4.3.

* src/math_funcs.cc, src/lineshapes.cc, src/legendre.cc,
src/m_general.cc, src/m_scatrte.cc, src/m_abs_lookup.cc,
src/m_jacobian.cc, src/m_montecarlo.cc, src/m_batch.cc,
src/montecarlo.cc, src/ppath.cc, src/rng.cc, src/test_matpack.cc,
src/test_integration.cc, src/matpackI.cc, src/methods_aux.cc,
src/geomag_calc.cc, src/m_atmosphere.cc, src/interpolation.cc,
src/binio.cc, src/gzstream.cc, src/main.cc, src/test_sparse.cc,
src/test_tensor.cc, src/test_interpolation.cc, src/test_sorting.cc:

Introduce explicit casts in several places.

* src/lineshapes.cc:

Add parenthesis in several places to make clear that + has precedence
over the shift operator <<.

* src/xml_io.cc (xml_set_stream_precision), src/methods_aux.cc,
src/abs_species_tags.cc:

Make variable 'precision' int instead of Index.

* src/nc_io.{cc,h}: Add function nc_get_data_long.

* src/nc_io_array_types.cc:

Change dimension types from int to long to match the matpack
dimensions.

* src/absorption.cc: Changed type of nChar from Numeric to Index.

* src/jacobian.cc (get_perturbation_limit):

Make inc Numeric instead of Index as it's anyway only used if FP
multiplications.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@5941 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/geomag_calc.cc
@@ -93,23 +93,23 @@ void magfield_nk(  // Output
 
 
           // Calculating the radial (upward) component of the magnetic field.
-          B_r += pow(l + 2, EARTH_RADIUS / r) * (l + 1) * 
-            ((M(j,0) + Ny * M(j,2)) * cos(m * Phi) 
-             + (M(j,1) + Ny * M(j,3)) * sin(m * Phi)) 
+          B_r += pow((Numeric)(l + 2), EARTH_RADIUS / r) * (Numeric)(l + 1) * 
+            ((M(j,0) + (Numeric)Ny * M(j,2)) * cos((Numeric)m * Phi) 
+             + (M(j,1) + (Numeric)Ny * M(j,3)) * sin((Numeric)m * Phi)) 
             * P_lm;
 
           // Calculating the latitudinal (southward) component of the magnetic field. 
           B_th += pow(l + 2, EARTH_RADIUS / r) * 
-            ((M(j,0) + Ny * M(j,2)) * cos(m * Phi) 
-             + (M(j,1) + Ny * M(j,3)) * sin(m * Phi)) *
+            ((M(j,0) + (Numeric)Ny * M(j,2)) * cos((Numeric)m * Phi) 
+             + (M(j,1) + (Numeric)Ny * M(j,3)) * sin((Numeric)m * Phi)) *
             dP_lm * sin(Theta);
 
 
 
           // Calculating the longitudinal (eastward) component of the magnetic field.
-          B_ph += pow(l + 2, EARTH_RADIUS / r) * m *
-            ((M(j,0) + Ny * M(j,2)) * sin(m * Phi) 
-             - (M(j,1) + Ny * M(j,3)) * cos(m * Phi)) *
+          B_ph += pow(l + 2, EARTH_RADIUS / r) * (Numeric)m *
+            ((M(j,0) + (Numeric)Ny * M(j,2)) * sin((Numeric)m * Phi) 
+             - (M(j,1) + (Numeric)Ny * M(j,3)) * cos((Numeric)m * Phi)) *
             P_lm / sin(Theta);
